Fixed test

+++ test/Utility/ModelTest.cpp
@@ -33,7 +33,7 @@ void testModel()
 	avrgBaseNormal += decltype(oldSFAVert.normal)(dbglNormal[0], dbglNormal[1], dbglNormal[2]);
     }
     avrgBaseNormal.normalize();
-    assert(dbgl::isSimilar((oldSFAVert.normal - avrgBaseNormal).norm(), 0, 0.0001f));
+    assert(dbgl::isSimilar((double)(oldSFAVert.normal - avrgBaseNormal).norm(), 0.0, 0.0001));
 
     // Rotate by 90° around y
     Eigen::AngleAxis<double> aa(dbgl::pi_2(), Eigen::Vector3d(0, 1, 0));
@@ -41,7 +41,7 @@ void testModel()
 
     // Check again
     auto newSFAVert = model.getVertex(0);
-    assert(dbgl::isSimilar((aa.toRotationMatrix() * oldSFAVert.normal - newSFAVert.normal).norm(), 0, 0.0001f));
+    assert(dbgl::isSimilar((double)(aa.toRotationMatrix() * oldSFAVert.normal - newSFAVert.normal).norm(), 0.0, 0.0001));
 
     avrgBaseNormal = decltype(oldSFAVert.normal)(0, 0, 0);
     for(auto i : newSFAVert.baseVertices)
@@ -50,5 +50,5 @@ void testModel()
 	avrgBaseNormal += decltype(newSFAVert.normal)(dbglNormal[0], dbglNormal[1], dbglNormal[2]);
     }
     avrgBaseNormal.normalize();
-    assert(dbgl::isSimilar((newSFAVert.normal - avrgBaseNormal).norm(), 0, 0.0001f));
+    assert(dbgl::isSimilar((double)(newSFAVert.normal - avrgBaseNormal).norm(), 0.0, 0.0001));
 }
